add a new `Search Results` page that can be used to preview search results
New `Search Results` page (`/ui-search`) can be used to preview up to 100
search results. This page can be accessed by clicking a link after
the list of search matches or by pressing `Enter` when `Search for files`
input is in focus and results are loaded.

To implement this feature we now cache full entries metadata, not only path.